--- config/runtime_metadata.py
@@ -1,396 +1,0 @@
-"""Safe read-only runtime metadata for sessions and sandboxed agent tools.
-
-Populated at session init so agents can answer introspection questions
-(e.g. OpenSRE version) without shelling out. Subprocess remains blocked in
-the Python execution sandbox; this is the preferred alternative.
-"""
-
-from __future__ import annotations
-
-import datetime as _dt
-import os
-import re
-import shutil
-import socket
-import sys
-import tempfile
-import time as _time
-from dataclasses import dataclass
-from pathlib import Path
-from typing import Any
-
-from config.config import get_environment
-from config.version import get_opensre_version
-
-# Reserved key merged into ``execute_python_code`` inputs (never overwrite user keys).
-RUNTIME_INPUTS_KEY = "opensre_runtime"
-
-_RELEASE_TAG_PATTERN = re.compile(r"^v\d+\.\d+(\.\d+){2,}$")
-
-
-def _resolve_gitdir(candidate: Path) -> Path | None:
-    """Return the git directory for ``candidate`` (``.git``), or ``None``.
-
-    Handles both a normal checkout (``.git`` is a directory) and a linked
-    worktree / submodule (``.git`` is a file with a ``gitdir: <path>`` line).
-    """
-    if candidate.is_dir():
-        return candidate
-    if not candidate.is_file():
-        return None
-    try:
-        content = candidate.read_text(encoding="utf-8")
-    except OSError:
-        return None
-    for line in content.splitlines():
-        stripped = line.strip()
-        if not stripped.startswith("gitdir:"):
-            continue
-        target = Path(stripped[len("gitdir:") :].strip())
-        if not target.is_absolute():
-            target = (candidate.parent / target).resolve()
-        return target if target.is_dir() else None
-    return None
-
-
-@dataclass(frozen=True)
-class _GitLayout:
-    """Per-worktree gitdir plus the shared common gitdir.
-
-    In a standard checkout the two are the same directory. In a linked worktree
-    (``git worktree add``), ``HEAD`` is per-worktree but ``refs/``, ``packed-refs``,
-    and tags live in the primary repo's gitdir named by the worktree's
-    ``commondir`` marker file.
-    """
-
-    gitdir: Path
-    commondir: Path
-
-
-def _resolve_commondir(gitdir: Path) -> Path:
-    """Return the shared common gitdir for ``gitdir``.
-
-    Standard checkouts have no ``commondir`` marker; the gitdir is its own
-    common dir. Linked worktrees carry a ``commondir`` file with a path
-    (relative to the per-worktree gitdir) to the primary repo's gitdir.
-    """
-    marker = gitdir / "commondir"
-    if not marker.is_file():
-        return gitdir
-    try:
-        content = marker.read_text(encoding="utf-8").strip()
-    except OSError:
-        return gitdir
-    if not content:
-        return gitdir
-    target = Path(content)
-    if not target.is_absolute():
-        target = (gitdir / target).resolve()
-    return target if target.is_dir() else gitdir
-
-
-def _find_git_layout() -> _GitLayout | None:
-    """Walk up from this file to the enclosing repo's git layout."""
-    here = Path(__file__).resolve().parent
-    while here.parent != here:
-        gitdir = _resolve_gitdir(here / ".git")
-        if gitdir is not None:
-            return _GitLayout(gitdir=gitdir, commondir=_resolve_commondir(gitdir))
-        here = here.parent
-    return None
-
-
-def _read_packed_refs(commondir: Path) -> dict[str, str]:
-    """Parse ``<commondir>/packed-refs`` into a ``{ref_name: sha}`` map.
-
-    After ``git pack-refs`` the loose files under ``refs/`` disappear and both
-    branch heads and tag refs live only here. Peeled tag lines (``^<sha>``) are
-    ignored: the non-peeled line already holds the tag object's sha which is
-    enough for a build marker.
-    """
-    packed = commondir / "packed-refs"
-    if not packed.is_file():
-        return {}
-    refs: dict[str, str] = {}
-    try:
-        content = packed.read_text(encoding="utf-8")
-    except OSError:
-        return {}
-    for raw in content.splitlines():
-        line = raw.strip()
-        if not line or line.startswith(("#", "^")):
-            continue
-        sha, _, name = line.partition(" ")
-        if sha and name:
-            refs[name] = sha
-    return refs
-
-
-def _read_ref_sha(layout: _GitLayout, ref_name: str) -> str | None:
-    """Resolve ``ref_name`` (e.g. ``refs/heads/main``) via loose files + packed-refs.
-
-    Per-worktree refs (bisect/HEAD-like) may live under the worktree gitdir,
-    so it's tried first; branches and tags live in the commondir.
-    """
-    for base in (layout.gitdir, layout.commondir):
-        loose = base / ref_name
-        if loose.is_file():
-            return loose.read_text(encoding="utf-8").strip() or None
-    return _read_packed_refs(layout.commondir).get(ref_name)
-
-
-def _read_git_head_sha(layout: _GitLayout) -> str | None:
-    """Short SHA the working tree currently points at, or ``None``."""
-    head_file = layout.gitdir / "HEAD"
-    if not head_file.is_file():
-        return None
-    head = head_file.read_text(encoding="utf-8").strip()
-    if not head.startswith("ref: "):
-        return head[:7] or None
-    sha = _read_ref_sha(layout, head[len("ref: ") :].strip())
-    return sha[:7] if sha else None
-
-
-def _release_tag_sort_key(name: str) -> tuple[int, ...] | None:
-    """Numeric tuple for a ``v0.1.YYYY.M.D`` tag; ``None`` if not all-numeric.
-
-    Numeric sort so ``v0.1.2026.10.1`` outranks ``v0.1.2026.9.30`` — a
-    lexicographic sort would pick the older tag because ``'9' > '1'`` as ASCII.
-    """
-    parts = name.removeprefix("v").split(".")
-    try:
-        return tuple(int(part) for part in parts)
-    except ValueError:
-        return None
-
-
-def _iter_release_tag_names(commondir: Path) -> set[str]:
-    """Release tag names, from loose refs and from ``packed-refs`` combined."""
-    names: set[str] = set()
-    tags_dir = commondir / "refs" / "tags"
-    if tags_dir.is_dir():
-        names.update(entry.name for entry in tags_dir.iterdir())
-    for ref_name in _read_packed_refs(commondir):
-        if ref_name.startswith("refs/tags/"):
-            names.add(ref_name[len("refs/tags/") :])
-    return names
-
-
-def _read_latest_release_tag(commondir: Path) -> str | None:
-    """Highest release tag (loose + packed) by numeric ordering."""
-    ranked: list[tuple[tuple[int, ...], str]] = []
-    for name in _iter_release_tag_names(commondir):
-        if not _RELEASE_TAG_PATTERN.match(name):
-            continue
-        key = _release_tag_sort_key(name)
-        if key is not None:
-            ranked.append((key, name))
-    if not ranked:
-        return None
-    ranked.sort(reverse=True)
-    return ranked[0][1]
-
-
-def _detect_build_info() -> str:
-    """Human-readable build marker: ``""`` for wheels, ``dev, <tag> @ <sha>`` for checkouts."""
-    layout = _find_git_layout()
-    if layout is None:
-        return ""
-    tag = _read_latest_release_tag(layout.commondir)
-    sha = _read_git_head_sha(layout)
-    if tag and sha:
-        return f"dev, {tag} @ {sha}"
-    if tag:
-        return f"dev, {tag}"
-    if sha:
-        return f"dev, @ {sha}"
-    return "dev"
-
-
-# Monotonic snapshot at import — anchor for uptime deltas that resist wall-clock skew.
-_PROCESS_START_MONOTONIC = _time.monotonic()
-
-# Tools the LLM commonly reflex-shells for. Presence-only surfaces so the agent
-# can answer without invoking ``--version``, which the sandbox blocks.
-_TOOLS_TO_PROBE = ("kubectl", "helm", "docker", "git", "python", "python3")
-
-
-_LOCALTIME_LINK = Path("/etc/localtime")
-
-
-def _local_tz_name() -> str:
-    """Best-effort local timezone name — IANA (``Europe/Berlin``) when possible.
-
-    Reads the ``/etc/localtime`` symlink target on macOS/Linux — the standard
-    way the OS advertises which zone it's set to. Falls back to
-    ``time.tzname`` short codes (``CET``, ``BST``) if the symlink can't be
-    resolved (Windows, unusual OS config), and finally to ``UTC``.
-    """
-    try:
-        if _LOCALTIME_LINK.is_symlink():
-            target = os.readlink(_LOCALTIME_LINK)
-            marker = "zoneinfo/"
-            idx = target.rfind(marker)
-            if idx >= 0:
-                iana = target[idx + len(marker) :]
-                if iana:
-                    return iana
-    except OSError:
-        # Unreadable /etc/localtime: fall back to time.tzname/UTC below.
-        pass
-    return _time.tzname[0] if _time.tzname else "UTC"
-
-
-def _python_version_string() -> str:
-    """Interpreter version as ``major.minor.patch`` from :mod:`sys` (no subprocess)."""
-    info = sys.version_info
-    return f"{info.major}.{info.minor}.{info.micro}"
-
-
-def _installed_tools() -> dict[str, str]:
-    """Map each probed tool name to its ``PATH`` location (empty if absent).
-
-    :func:`shutil.which` walks ``PATH`` in pure Python — no subprocess. Version
-    strings would require invoking the binary and are intentionally omitted;
-    presence is what the agent needs to stop reflex-shelling for ``--version``.
-    """
-    return {tool: shutil.which(tool) or "" for tool in _TOOLS_TO_PROBE}
-
-
-_HOSTNAME_FILE = Path("/etc/hostname")
-
-
-def _pod_hostname() -> str:
-    """Hostname via file read, no subprocess.
-
-    ``/etc/hostname`` holds the pod name inside Kubernetes containers, which is
-    the value SRE questions ("which pod am I in?") actually want. Falls back to
-    :func:`socket.gethostname` on hosts without the file (macOS, some distros).
-    """
-    try:
-        if _HOSTNAME_FILE.is_file():
-            name = _HOSTNAME_FILE.read_text(encoding="utf-8").strip()
-            if name:
-                return name
-    except OSError:
-        # Unreadable /etc/hostname: fall back to the socket API below.
-        pass
-    try:
-        return socket.gethostname()
-    except OSError:
-        return ""
-
-
-def _disk_memory_facts() -> dict[str, Any]:
-    """Live disk and memory readings via psutil (no ``df``/``free``/``top``).
-
-    Degrades to an empty dict if psutil misbehaves on an exotic platform —
-    the facts are then simply absent rather than crashing prompt assembly.
-    """
-    try:
-        import psutil
-
-        disk = psutil.disk_usage("/")
-        memory = psutil.virtual_memory()
-    except Exception:
-        return {}
-    gib = 1024**3
-    return {
-        "disk_used_percent": round(disk.percent, 1),
-        "disk_free_gb": round(disk.free / gib, 1),
-        "memory_used_percent": round(memory.percent, 1),
-        "memory_available_gb": round(memory.available / gib, 1),
-    }
-
-
-def _kubeconfig_path() -> str:
-    """Effective ``kubeconfig`` path from env, or the default under ``~/.kube``.
-
-    Kept as a session-static fact so the agent can answer "which cluster
-    config is loaded" without shelling to ``kubectl config view``.
-    """
-    override = (os.environ.get("KUBECONFIG") or "").strip()
-    if override:
-        # ``KUBECONFIG`` may be a ``:``-separated list; the first entry wins.
-        first = override.split(os.pathsep, 1)[0]
-        if first:
-            return first
-    default = Path.home() / ".kube" / "config"
-    return str(default) if default.is_file() else ""
-
-
-def build_runtime_metadata() -> dict[str, Any]:
-    """Session-lifetime read-only runtime facts.
-
-    Keys are stable for prompts and sandbox ``inputs`` and safe to cache at
-    session bootstrap (nothing here changes turn to turn):
-
-    - ``opensre_version`` — package version via ``importlib.metadata``.
-    - ``opensre_build`` — ``""`` in released wheels; ``dev, v0.1.YYYY.M.D @ SHA``
-      in a git checkout so the LLM can quote the exact build in local dev.
-    - ``runtime_env`` — ``OPENSRE_ENV`` env var, else the app environment name.
-    - ``tz_name`` — local timezone name (rarely changes mid-session).
-    - ``python_version`` — interpreter version from :data:`sys.version_info`.
-    - ``pid`` / ``ppid`` — this process and its parent from :mod:`os`.
-    - ``tools`` — probed tool paths (``kubectl``, ``helm``, ``docker``, ``git``, …).
-    - ``kubeconfig`` — effective kubeconfig path (``KUBECONFIG`` or ``~/.kube/config``).
-    - ``hostname`` — from ``/etc/hostname`` (the pod name in Kubernetes) or
-      :func:`socket.gethostname`, never the ``hostname`` binary.
-    - ``scratchpad_dir`` — the temp directory scripts may write to.
-
-    Live values that must NOT be cached (current time, uptime, disk, memory)
-    come from :func:`capture_runtime_facts` at each render/sandbox call.
-    """
-    env_override = (os.environ.get("OPENSRE_ENV") or "").strip()
-    return {
-        "opensre_version": get_opensre_version(),
-        "opensre_build": _detect_build_info(),
-        "runtime_env": env_override or get_environment().value,
-        "tz_name": _local_tz_name(),
-        "python_version": _python_version_string(),
-        "pid": os.getpid(),
-        "ppid": os.getppid(),
-        "tools": _installed_tools(),
-        "kubeconfig": _kubeconfig_path(),
-        "hostname": _pod_hostname(),
-        "scratchpad_dir": tempfile.gettempdir(),
-    }
-
-
-def capture_runtime_facts(*, metadata: dict[str, Any] | None = None) -> dict[str, Any]:
-    """Session metadata plus fresh live facts (time, uptime, disk, memory).
-
-    Read once per prompt render or sandbox invocation so time doesn't lie. Pass
-    ``metadata`` (typically ``session.runtime_metadata``) to avoid re-running
-    the git+importlib probe every call.
-    """
-    facts = dict(metadata or build_runtime_metadata())
-    facts["now_iso"] = _dt.datetime.now().astimezone().isoformat(timespec="seconds")
-    facts["uptime_seconds"] = round(_time.monotonic() - _PROCESS_START_MONOTONIC, 3)
-    facts.update(_disk_memory_facts())
-    return facts
-
-
-def merge_runtime_into_inputs(
-    inputs: dict[str, Any] | None,
-    *,
-    metadata: dict[str, Any] | None = None,
-) -> dict[str, Any]:
-    """Copy ``inputs`` and inject runtime facts under :data:`RUNTIME_INPUTS_KEY`.
-
-    Never overwrites an existing ``opensre_runtime`` key supplied by the caller.
-    Facts are captured live via :func:`capture_runtime_facts` so ``now_iso`` is
-    fresh for each sandbox invocation.
-    """
-    merged: dict[str, Any] = dict(inputs or {})
-    if RUNTIME_INPUTS_KEY not in merged:
-        merged[RUNTIME_INPUTS_KEY] = capture_runtime_facts(metadata=metadata)
-    return merged
-
-
-__all__ = [
-    "RUNTIME_INPUTS_KEY",
-    "build_runtime_metadata",
-    "capture_runtime_facts",
-    "merge_runtime_into_inputs",
-]
